+++ crates/telio-model/src/api_config.rs
@@ -1,7 +1,7 @@
 //! Object descriptions of various
 //! telio configurable features via API
 
-use std::{collections::HashSet, fmt};
+use std::{collections::HashSet, fmt, time::Duration};
 
 use num_enum::{IntoPrimitive, TryFromPrimitive};
 use serde::{de::IntoDeserializer, Deserialize, Deserializer, Serialize};
@@ -94,13 +94,32 @@ pub struct FeatureNurse {
     /// The unique identifier of the device, used for meshnet ID
     pub fingerprint: String,
     /// Heartbeat interval in seconds. Default value is 3600.
+    #[serde(default = "FeatureNurse::get_default_heartbeat_interval")]
     pub heartbeat_interval: Option<u32>,
-    /// Initial heartbeat interval in seconds. Default value is None.
+    /// Initial heartbeat interval in seconds. Default value is 300.
+    #[serde(default = "FeatureNurse::get_default_initial_heartbeat_interval")]
     pub initial_heartbeat_interval: Option<u32>,
     /// QoS configuration for Nurse
     pub qos: Option<FeatureQoS>,
     /// Enable/disable collecting nat type
     pub enable_nat_type_collection: Option<bool>,
+}
+
+impl FeatureNurse {
+    /// One hour
+    pub const DEFAULT_HEARTBEAT_INTERVAL: Duration = Duration::from_secs(3600);
+    /// 5 minutes
+    pub const DEFAULT_INITIAL_HEARTBEAT_INTERVAL: Duration = Duration::from_secs(300);
+
+    fn get_default_heartbeat_interval() -> Option<u32> {
+        // 3600 seconds
+        Some(Self::DEFAULT_HEARTBEAT_INTERVAL.as_secs() as u32)
+    }
+
+    fn get_default_initial_heartbeat_interval() -> Option<u32> {
+        // 300 seconds
+        Some(Self::DEFAULT_INITIAL_HEARTBEAT_INTERVAL.as_secs() as u32)
+    }
 }
 
 /// Configurable features for Lana module
@@ -493,8 +512,8 @@ mod tests {
         },
         nurse: Some(FeatureNurse {
             fingerprint: "fingerprint_test".to_string(),
-            heartbeat_interval: None,
-            initial_heartbeat_interval: None,
+            heartbeat_interval: Some(3600),
+            initial_heartbeat_interval: Some(300),
             qos: None,
             enable_nat_type_collection: None,
         }),
@@ -549,8 +568,8 @@ mod tests {
         },
         nurse: Some(FeatureNurse {
             fingerprint: "fingerprint_test".to_string(),
-            heartbeat_interval: None,
-            initial_heartbeat_interval: None,
+            heartbeat_interval: Some(3600),
+            initial_heartbeat_interval: Some(300),
             qos: None,
             enable_nat_type_collection: None,
         }),
@@ -705,6 +724,7 @@ mod tests {
             "nurse": {
                 "fingerprint": "fingerprint_test",
                 "heartbeat_interval": 3600,
+                "initial_heartbeat_interval": 300,
                 "qos": {
                     "rtt_interval": 3600,
                     "rtt_tries": 5,
@@ -734,7 +754,7 @@ mod tests {
             nurse: Some(FeatureNurse {
                 fingerprint: String::from("fingerprint_test"),
                 heartbeat_interval: Some(3600),
-                initial_heartbeat_interval: None,
+                initial_heartbeat_interval: Some(300),
                 qos: Some(FeatureQoS {
                     rtt_interval: Some(3600),
                     rtt_tries: Some(5),
@@ -763,8 +783,8 @@ mod tests {
             wireguard: Default::default(),
             nurse: Some(FeatureNurse {
                 fingerprint: String::from("fingerprint_test"),
-                heartbeat_interval: None,
-                initial_heartbeat_interval: None,
+                heartbeat_interval: Some(3600),
+                initial_heartbeat_interval: Some(300),
                 qos: Some(FeatureQoS {
                     rtt_interval: None,
                     rtt_tries: None,
@@ -793,8 +813,8 @@ mod tests {
             wireguard: Default::default(),
             nurse: Some(FeatureNurse {
                 fingerprint: String::from("fingerprint_test"),
-                heartbeat_interval: None,
-                initial_heartbeat_interval: None,
+                heartbeat_interval: Some(3600),
+                initial_heartbeat_interval: Some(300),
                 qos: None,
                 enable_nat_type_collection: None,
             }),